Validation of file extensions
A dictionary containing valid file extensions for picutres or movies was created.
Only the files with the extensions in the dictionary are going to be renamed.

Additionally a previous bug was fixed, the program was not renaming the files with a valid extension written in capital letters.
The suffix of the file was being compared with the dictionary keys in lower case, so the program was not recognizing the valid extensions.